+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ latest
 ------
 
 * Display graph in browser rather than using Graphviz.
+* Render nodes as relative imports.
 
 1.2 (2025-10-16)
 ----------------